feat: make stuff consistent across the whole forge

+++ include/bfs_graph_matrix.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-optional<vector<int>> bfs(const vector<vector<int64_t>>& g, int source, int sink);
+optional<vector<int>> bfs(const vector<vector<int>>& g, int source, int sink);
