+++ include/bt_bfs.h
@@ -4,4 +4,4 @@ using namespace std;
 
 struct Node { int val; Node* left; Node* right; };
 
-bool bfs(Node* root, int64_t needle);
+bool bfs(Node* root, int needle);
