+++ include/coin_change.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t coin_change(int64_t amount, const vector<int>& coins);
+int coin_change(int amount, const vector<int>& coins);
