+++ include/combinations.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-vector<vector<int64_t>> combine(int64_t n, int64_t k);
+vector<vector<int>> combine(int n, int k);
