+++ include/find_min_rotated_sorted.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t find_min_rotated(const vector<int>& a);
+int find_min_rotated(const vector<int>& a);
