+++ include/kadanes_algorithm.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t kadane(const vector<int>& a);
+int kadane(const vector<int>& a);
