+++ include/knapsack.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t knapsack(int64_t capacity, const vector<int>& weights, const vector<int>& values);
+int knapsack(int capacity, const vector<int>& weights, const vector<int>& values);
