+++ include/longest_common_subsequence.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t longest_common_subsequence(const string& a, const string& b);
+int longest_common_subsequence(const string& a, const string& b);
