+++ include/map.h
@@ -3,8 +3,8 @@
 using namespace std;
 
 struct Map {
-    void set(const string& k, int64_t v);
-    optional<int64_t> get(const string& k);
+    void set(const string& k, int v);
+    optional<int> get(const string& k);
     void remove(const string& k);
     int size();
 };
