+++ include/subsets.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-vector<vector<int64_t>> subsets(const vector<int64_t>& a);
+vector<vector<int>> subsets(const vector<int>& a);
