+++ include/two_crystal_balls.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int64_t two_crystal_balls(const vector<int>& a);
+int two_crystal_balls(const vector<int>& a);
