+++ include/two_sum_sorted.h
@@ -2,4 +2,4 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-vector<int> two_sum(const vector<int>& a, int64_t k);
+vector<int> two_sum(const vector<int>& a, int k);
