+++ tests/dfs_on_bst.test.cpp
@@ -3,21 +3,21 @@
 int main() {
     { TEST("found_45");
         Node* _tree = new Node{20, new Node{10, new Node{5, nullptr, new Node{7, nullptr, nullptr}}, new Node{15, nullptr, nullptr}}, new Node{50, new Node{30, new Node{29, nullptr, nullptr}, new Node{45, nullptr, nullptr}}, new Node{100, nullptr, nullptr}}};
-        int64_t _needle = 45;
+        int _needle = 45;
         auto _r = dfs(_tree, _needle);
         CHECK(_r == true);
     }
 
     { TEST("found_7");
         Node* _tree = new Node{20, new Node{10, new Node{5, nullptr, new Node{7, nullptr, nullptr}}, new Node{15, nullptr, nullptr}}, new Node{50, new Node{30, new Node{29, nullptr, nullptr}, new Node{45, nullptr, nullptr}}, new Node{100, nullptr, nullptr}}};
-        int64_t _needle = 7;
+        int _needle = 7;
         auto _r = dfs(_tree, _needle);
         CHECK(_r == true);
     }
 
     { TEST("missing");
         Node* _tree = new Node{20, new Node{10, new Node{5, nullptr, new Node{7, nullptr, nullptr}}, new Node{15, nullptr, nullptr}}, new Node{50, new Node{30, new Node{29, nullptr, nullptr}, new Node{45, nullptr, nullptr}}, new Node{100, nullptr, nullptr}}};
-        int64_t _needle = 69;
+        int _needle = 69;
         auto _r = dfs(_tree, _needle);
         CHECK(_r == false);
     }
