+++ tests/merge_ranges.test.cpp
@@ -2,33 +2,33 @@
 
 int main() {
     { TEST("overlap_and_disjoint");
-        vector<pair<int64_t,int64_t>> _pairs = {{1,3}, {2,6}, {8,10}, {15,18}};
+        vector<pair<int,int>> _pairs = {{1,3}, {2,6}, {8,10}, {15,18}};
         auto _r = merge_ranges(_pairs);
-        CHECK(_r == vector<pair<int64_t,int64_t>>{{1,6}, {8,10}, {15,18}});
+        CHECK(_r == vector<pair<int,int>>{{1,6}, {8,10}, {15,18}});
     }
 
     { TEST("adjacent_touch");
-        vector<pair<int64_t,int64_t>> _pairs = {{1,4}, {4,5}};
+        vector<pair<int,int>> _pairs = {{1,4}, {4,5}};
         auto _r = merge_ranges(_pairs);
-        CHECK(_r == vector<pair<int64_t,int64_t>>{{1,5}});
+        CHECK(_r == vector<pair<int,int>>{{1,5}});
     }
 
     { TEST("disjoint");
-        vector<pair<int64_t,int64_t>> _pairs = {{1,2}, {3,4}};
+        vector<pair<int,int>> _pairs = {{1,2}, {3,4}};
         auto _r = merge_ranges(_pairs);
-        CHECK(_r == vector<pair<int64_t,int64_t>>{{1,2}, {3,4}});
+        CHECK(_r == vector<pair<int,int>>{{1,2}, {3,4}});
     }
 
     { TEST("empty_input");
-        vector<pair<int64_t,int64_t>> _pairs = {};
+        vector<pair<int,int>> _pairs = {};
         auto _r = merge_ranges(_pairs);
-        CHECK(_r == vector<pair<int64_t,int64_t>>{});
+        CHECK(_r == vector<pair<int,int>>{});
     }
 
     { TEST("fully_contained");
-        vector<pair<int64_t,int64_t>> _pairs = {{1,10}, {2,3}, {4,5}};
+        vector<pair<int,int>> _pairs = {{1,10}, {2,3}, {4,5}};
         auto _r = merge_ranges(_pairs);
-        CHECK(_r == vector<pair<int64_t,int64_t>>{{1,10}});
+        CHECK(_r == vector<pair<int,int>>{{1,10}});
     }
 
     return 0;
